+++ app/models/spree/order_decorator.rb
@@ -1,42 +1,37 @@
 require 'logger'
 
 Spree::Order.class_eval do
+
   has_one :avalara_transaction, dependent: :destroy
 
-  self.state_machine.after_transition to: :complete,
-                                      do: :avalara_capture_finalize,
-                                      if: :avalara_eligible
+  self.state_machine.before_transition :to => :canceled,
+                                      :do => :cancel_avalara,
+                                      :if => :avalara_eligible?
 
- self.state_machine.before_transition to: :canceled,
-                                      do: :cancel_status,
-                                      if: :avalara_eligible
-
-  def avalara_eligible
+  def avalara_eligible?
     Spree::Config.avatax_iseligible
   end
 
   def avalara_lookup
     logger.debug 'avalara lookup'
-    create_avalara_transaction
+    create_avalara_transaction if avalara_transaction.nil?
     :lookup_avatax
   end
 
-  def cancel_status
+  def cancel_avalara
     return nil unless avalara_transaction.present?
-    avalara_transaction.check_status(self)
+    self.avalara_transaction.cancel_order
   end
 
   def avalara_capture
     logger.debug 'avalara capture'
-
     begin
-      create_avalara_transaction
+      create_avalara_transaction if avalara_transaction.nil?
       line_items.reload
 
-      @rtn_tax = avalara_transaction.commit_avatax(line_items, self, number.to_s, Date.today.strftime('%F'), 'SalesInvoice')
+      @rtn_tax = self.avalara_transaction.commit_avatax('SalesInvoice')
 
-      logger.info 'tax amount'
-      logger.debug @rtn_tax
+      logger.info_and_debug('tax amount', @rtn_tax)
       @rtn_tax
     rescue => e
       logger.debug e
@@ -47,12 +42,11 @@ Spree::Order.class_eval do
   def avalara_capture_finalize
     logger.debug 'avalara capture finalize'
     begin
-      create_avalara_transaction
+      create_avalara_transaction if avalara_transaction.nil?
       line_items.reload
-      @rtn_tax = avalara_transaction.commit_avatax_final(line_items, self, number.to_s, Date.today.strftime('%F'), 'SalesInvoice')
+      @rtn_tax = avalara_transaction.commit_avatax_final('SalesInvoice')
 
-      logger.info 'tax amount'
-      logger.debug @rtn_tax
+      logger.info_and_debug('tax amount', @rtn_tax)
       @rtn_tax
     rescue => e
       logger.debug e
@@ -62,8 +56,8 @@ Spree::Order.class_eval do
 
   def avatax_cache_key
     key = ['Spree::Order']
-    key << number
-    key << promo_total
+    key << self.number
+    key << self.promo_total
     key.join('-')
   end