+++ app/models/spree/refund_decorator.rb
@@ -1,14 +1,13 @@
- require 'logger'
- REFUND_LOGGER = AvataxHelper::AvataxLog.new('refund', 'refund class')
+require 'logger'
+REFUND_LOGGER = AvataxHelper::AvataxLog.new('refund', 'refund class')
 
- Spree::Refund.class_eval do
+Spree::Refund.class_eval do
   REFUND_LOGGER.info('start refund processing')
 
-  has_one :avalara_transaction, dependent: :destroy
-  after_save :assign_avalara_transaction, on: :create
-  after_save :avalara_capture_finalize, on: :update
+  has_one :avalara_transaction
+  after_create :avalara_capture_finalize, if: :avalara_eligible?
 
-  def avalara_eligible
+  def avalara_eligible?
     Spree::Config.avatax_iseligible
   end
 
@@ -18,66 +17,19 @@
     :lookup_avatax
   end
 
-  def avalara_capture
-    if self.reimbursement_id.nil?
-      REFUND_LOGGER.debug 'avalara capture refund'
-      begin
-        avalara_transaction_refund = Spree::AvalaraTransaction.find_by_refund_id(self.id)
-        @rtn_tax = avalara_transaction_refund.commit_avatax(payment.order.line_items, payment.order, payment.order.number.to_s + '.' + self.id.to_s, avalara_transaction_refund.order.completed_at.strftime('%F'), 'ReturnInvoice')
-
-        REFUND_LOGGER.info 'tax amount'
-        REFUND_LOGGER.debug @rtn_tax
-
-        @rtn_tax
-      rescue => e
-        REFUND_LOGGER.debug e
-        REFUND_LOGGER.debug 'error in avalara capture refund'
-      end
-    end
-  end
-
   def avalara_capture_finalize
-    if self.reimbursement_id.nil?
-      REFUND_LOGGER.debug 'avalara capture refund avalara_capture_finalize'
-      begin
-        avalara_transaction_refund = Spree::AvalaraTransaction.find_by_refund_id(self.id)
-        @rtn_tax = avalara_transaction_refund.commit_avatax_final(payment.order.line_items, payment.order, payment.order.number.to_s + '.' + self.id.to_s, avalara_transaction_refund.order.completed_at.strftime('%F'), 'ReturnInvoice')
+    REFUND_LOGGER.debug 'avalara capture refund avalara_capture_finalize'
+    begin
+      avalara_transaction_refund = Spree::AvalaraTransaction.find_by_order_id(self.payment.order.id)
 
-        REFUND_LOGGER.info 'tax amount'
-        REFUND_LOGGER.debug @rtn_tax
+      @rtn_tax = avalara_transaction_refund.commit_avatax_final('ReturnInvoice', self.id)
 
-        @rtn_tax
-      rescue => e
-        REFUND_LOGGER.debug e
-        REFUND_LOGGER.debug 'error in avalara capture refund finalize'
-      end
+      REFUND_LOGGER.info 'tax amount'
+      REFUND_LOGGER.debug @rtn_tax
+      @rtn_tax
+    rescue => e
+      REFUND_LOGGER.debug e
+      REFUND_LOGGER.debug 'error in avalara capture refund finalize'
     end
-  end
-
-  def create_avalara_transaction_refund
-    Spree::AvalaraTransaction.create(order_id: payment.order.id, refund_id: self.id)
-  end
-
-  def assign_avalara_transaction
-    if avalara_eligible
-      if self.payment.order.avalara_transaction.nil?
-        create_avalara_transaction_refund
-      else
-        Spree::AvalaraTransaction.find_by_order_id(payment.order.id).update_attributes(order_id: payment.order.id, refund_id: self.id)
-      end
-    end
-  end
-
-  def pre_tax_amount
-    unless payment.order.additional_tax_total.zero?
-      tax_amount = tax_percentage * amount
-      return amount - tax_amount
-    end
-  end
-
-  private
-
-  def tax_percentage
-    (payment.order.additional_tax_total / payment.order.total).to_f
   end
 end